requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
supabase==2.0.2
httpx>=0.24.0,<0.25.0
lightfm==1.17
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
shap==0.43.0
lime==0.2.0.1
alibi==0.9.4
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
psycopg2-binary==2.9.9
Pillow==9.5.0   # downgraded for alibi compatibility
python-decouple==3.8
